# Conflicts:
#	src/main/java/com/ec/entity/CartItemEntity.java
#	src/main/java/com/ec/entity/CategoryItemEntity.java

+++ src/main/java/com/ec/dao/OrderDao.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrderDao extends JpaRepository<OrderEntity,Integer> {
+public interface OrderDao extends JpaRepository<OrderEntity,Long> {
 }
